(fix): Reorder pathbuilder groups
Due to how the path-tree is assembled, groups acting as parents must
appear before their child fields in order for the hierarchy to be
reconstructed correctly.

However, the pathbuilder XML does not guarantee this order, and groups
may appear _after_ their children, thus breaking the tree.

Fix this by re-ordering the path list so that all groups appear at the
front, followed by all non-groups (sort-of).