(feat): Don't eagerly load pathbuilders
In the API constructor, we always load all available pathbuilders and
internalize them. In the case where a specific pathbuilder is configured
by direct assignment to `self.pathbuilders`, this internalization will
happen again, making the first run redundant.

This patch introduces an explicit init-method that can be called if all
pathbuilders are desired; if a (set of) specific pathbuilder(s) is set
at a later stage, this will still be handled by the existing
setter-logic.